Bug-fixes for pruner.yml
For the rolebinding task, serviceaccounts is meant to be a list but was
being set to a string. As a result the template tried to create a new
service account for each character.

The CronJob command now limits the pruning operation to a single
namespace.

Additionally, the rolebinding needs to be a clusterrolebinding because
'oc adm' requires cluster-wide permissions even when limiting its
operation to a single namespace.

Further, both cluster-reader (to list BuildConfigs) and
build-config-change-controller (to delete Builds) cluster roles need to
be included in the binding.

Finally, the name of the clusterrolebinding needs to include the
namespace to avoid collisions when run for multiple namespaces in the
cluster.

Signed-off-by: Tim Waugh <twaugh@redhat.com>